Move zipping of files for support purposes into its own folder
